requirements_m1.txt
Cython==0.29.35
Pillow==9.4.0
Pypubsub==4.0.3
h5py==3.8.0
imageio==2.31.1
nibabel==5.1.0
numpy==1.24.3
psutil==5.9.4
pyserial==3.5
python-gdcm==3.0.22
scikit-image==0.21.0
scipy==1.10.1
vtk==9.2.6
wxPython==4.2.1
torch==2.0.1
pyacvd==0.2.9
pynetdicom==2.0.2
